Changed mesh sampling for qrsm to sample a rectangle inscribed by the given ellipse.
Previously mesh sampling of the quadratic function sampled the
rectangular domain specified by the lo and hi bounds of the
uncertainty domain. That's fine for the inner ellipse but won't find
the extrema in the outer ellipse since it doesn't cover the domain of
the outer ellipse. This fixes that issue.